made a validate waypoint function, not sure i put it in the right place tho
